requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
statsmodels
scikit-learn
lifelines
streamlit>=1.31.0
altair<5.0.0
jinja2
networkx
GEOparse
requests
gseapy
matplotlib-venn
adjustText
pywebview
